editor: created a joint media form and component. improved media blocks styling and error texts. added setting cursor at the next block after uploading media block
