+++ frontend/packages/editor/src/embed-block.tsx
@@ -18,25 +18,17 @@ import {
   Check,
   ChevronDown,
   ExternalLink,
-  Form,
-  Input,
   Popover,
-  SizableText,
-  Spinner,
-  Tabs,
   Tooltip,
   XStack,
-  YStack,
-  useTheme,
 } from '@mintter/ui'
-import {useCallback, useEffect, useMemo, useState} from 'react'
+import {useCallback, useMemo} from 'react'
 import {ErrorBoundary} from 'react-error-boundary'
-import {RiArticleLine} from 'react-icons/ri'
 import {ListItem, Separator, YGroup} from 'tamagui'
-import {Block, BlockNoteEditor, HMBlockSchema, getBlockInfoFromPos} from '.'
+import {Block, BlockNoteEditor, HMBlockSchema} from '.'
 import {createReactBlockSpec} from './blocknote/react'
+import {DisplayComponentProps, MediaRender, MediaType} from './media-render'
 import {usePopoverState} from './use-popover-state'
-type LinkType = null | 'basic' | 'hypermedia'
 
 function EmbedError() {
   return <ErrorBlock message="Failed to load this Embedded document" />
@@ -45,7 +37,7 @@ function EmbedError() {
 export const EmbedBlock = createReactBlockSpec({
   type: 'embed',
   propSchema: {
-    ref: {
+    url: {
       default: '',
     },
     defaultOpen: {
@@ -68,134 +60,119 @@ export const EmbedBlock = createReactBlockSpec({
   }) => Render(block, editor),
 })
 
-type EmbedType = {
-  id: string
-  props: {
-    ref: string
-    display?: 'content' | 'card'
-  }
-  children: []
-  content: []
-  type: string
-}
-
-const boolRegex = new RegExp('true')
-
 const Render = (
   block: Block<HMBlockSchema>,
   editor: BlockNoteEditor<HMBlockSchema>,
 ) => {
-  const [selected, setSelected] = useState(false)
-  const tiptapEditor = editor._tiptapEditor
-  const selection = tiptapEditor.state.selection
-
-  useEffect(() => {
-    const selectedNode = getBlockInfoFromPos(
-      tiptapEditor.state.doc,
-      tiptapEditor.state.selection.from,
-    )
-    if (selectedNode && selectedNode.id) {
-      if (
-        selectedNode.id === block.id &&
-        selectedNode.startPos === selection.$anchor.pos
-      ) {
-        setSelected(true)
-      } else if (selectedNode.id !== block.id) {
-        setSelected(false)
+  const {queryClient} = useAppContext()
+  const gwUrl = useGatewayUrlStream()
+  const submitEmbed = async (
+    url: string,
+    assign: any,
+    setFileName: any,
+    setLoading: any,
+  ) => {
+    if (isPublicGatewayLink(url, gwUrl) || isHypermediaScheme(url)) {
+      const hmLink = normlizeHmId(url, gwUrl)
+      const newUrl = hmLink ? hmLink : url
+      assign({props: {url: newUrl}} as MediaType)
+      const cursorPosition = editor.getTextCursorPosition()
+      editor.focus()
+      if (cursorPosition.block.id === block.id) {
+        if (cursorPosition.nextBlock)
+          editor.setTextCursorPosition(cursorPosition.nextBlock, 'start')
+        else {
+          editor.insertBlocks(
+            [{type: 'paragraph', content: ''}],
+            block.id,
+            'after',
+          )
+          editor.setTextCursorPosition(
+            editor.getTextCursorPosition().nextBlock!,
+            'start',
+          )
+        }
       }
+    } else {
+      setLoading(true)
+      fetchWebLink(queryClient, url)
+        .then((res) => {
+          const fullHmId = hmIdWithVersion(
+            res?.hmId,
+            res?.hmVersion,
+            extractBlockRefOfUrl(url),
+          )
+          if (fullHmId) {
+            assign({props: {url: fullHmId}} as MediaType)
+            const cursorPosition = editor.getTextCursorPosition()
+            editor.focus()
+            if (cursorPosition.block.id === block.id) {
+              if (cursorPosition.nextBlock)
+                editor.setTextCursorPosition(cursorPosition.nextBlock, 'start')
+              else {
+                editor.insertBlocks(
+                  [{type: 'paragraph', content: ''}],
+                  block.id,
+                  'after',
+                )
+                editor.setTextCursorPosition(
+                  editor.getTextCursorPosition().nextBlock!,
+                  'start',
+                )
+              }
+            }
+          } else {
+            setFileName({
+              name: 'The provided url is not a hypermedia link',
+              color: 'red',
+            })
+          }
+          setLoading(false)
+        })
+        .catch((e) => {
+          setFileName({
+            name: 'The provided url is not a hypermedia link',
+            color: 'red',
+          })
+          setLoading(false)
+        })
     }
-  }, [selection, editor, block.id, tiptapEditor])
-
-  const assignEmbed = (newEmbed: EmbedType) => {
-    editor.updateBlock(block.id, {
-      props: {...block.props, ...newEmbed.props},
-    })
   }
-
-  const setSelection = (isSelected: boolean) => {
-    setSelected(isSelected)
-  }
-
   return (
-    <YStack
-      backgroundColor={selected ? '$color4' : '$color3'}
-      borderColor={selected ? '$color8' : 'transparent'}
-      borderWidth={2}
-      borderRadius="$4"
-      overflow="hidden"
-      hoverStyle={{
-        backgroundColor: '$color4',
-      }}
-    >
-      {block.props.ref ? (
-        <EmbedComponent
-          block={block}
-          editor={editor}
-          assign={assignEmbed}
-          selected={selected}
-        />
-      ) : editor.isEditable ? (
-        <EmbedForm
-          block={block}
-          assign={assignEmbed}
-          editor={editor}
-          selected={selected}
-        />
-      ) : (
-        <></>
-      )}
-    </YStack>
+    <MediaRender
+      block={block}
+      editor={editor}
+      mediaType="embed"
+      submit={submitEmbed}
+      DisplayComponent={display}
+      icon={<ExternalLink />}
+    />
   )
 }
 
-function EmbedComponent({
-  block,
-  editor,
-  assign,
-  selected,
-}: {
-  block: Block<HMBlockSchema>
-  editor: BlockNoteEditor<HMBlockSchema>
-  assign: any
-  selected: boolean
-}) {
+const display = ({block, assign}: DisplayComponentProps) => {
   return (
-    <YStack gap="$2" position="relative">
-      <YStack
-        backgroundColor={selected ? '$color4' : '$color3'}
-        borderColor={selected ? '$color8' : 'transparent'}
-        borderWidth={2}
-        borderRadius="$2"
-        overflow="hidden"
-        hoverStyle={{
-          backgroundColor: '$color4',
-        }}
-        // @ts-ignore
-        contentEditable={false}
-        className={block.type}
-        group="item"
-      >
-        <EmbedControl block={block} assign={assign} />
-        {block.props.ref && (
-          <ErrorBoundary FallbackComponent={EmbedError}>
-            <BlockContentEmbed
-              block={{
-                id: block.id,
-                type: 'embed',
-                text: ' ',
-                attributes: {
-                  childrenType: 'group',
-                  view: block.props.view,
-                },
-                annotations: [],
-                ref: block.props.ref,
-              }}
-              depth={1}
-            />
-          </ErrorBoundary>
-        )}
-      </YStack>
-    </YStack>
+    <>
+      <EmbedControl block={block} assign={assign} />
+      {block.props.url && (
+        <ErrorBoundary FallbackComponent={EmbedError}>
+          <BlockContentEmbed
+            block={{
+              id: block.id,
+              type: 'embed',
+              text: ' ',
+              attributes: {
+                childrenType: 'group',
+                view: block.props.view,
+              },
+              annotations: [],
+              ref: block.props.url,
+            }}
+            depth={1}
+          />
+        </ErrorBoundary>
+      )}
+    </>
   )
 }
 
@@ -207,11 +184,11 @@ function EmbedControl({
   assign: any
 }) {
   const hmId = useMemo(() => {
-    if (block.props.ref) {
-      return unpackHmId(block.props.ref)
+    if (block.props.url) {
+      return unpackHmId(block.props.url)
     }
     return null
-  }, [block.props.ref])
+  }, [block.props.url])
   const allowViewSwitcher = hmId?.type === 'd' && !hmId.blockRef
   const allowVersionSwitcher = hmId?.type === 'd'
   const openUrl = useOpenUrl()
@@ -219,7 +196,7 @@ function EmbedControl({
   const popoverViewState = usePopoverState()
   const popoverLatestState = usePopoverState()
 
-  let versionValue = block.props.ref.includes('&l') ? 'latest' : 'exact'
+  let versionValue = block.props.url.includes('&l') ? 'latest' : 'exact'
   let isVersionLatest = versionValue == 'latest'
 
   const handleViewSelect = useCallback((view: 'content' | 'card') => {
@@ -231,13 +208,13 @@ function EmbedControl({
 
   const handleVersionSelect = useCallback(
     (versionMode: 'exact' | 'latest') => {
-      let unpackedRef = unpackHmId(block.props.ref)
+      let unpackedRef = unpackHmId(block.props.url)
       return () => {
         popoverLatestState.onOpenChange(false)
         if (unpackedRef) {
           assign({
             props: {
-              ref: createHmDocLink({
+              url: createHmDocLink({
                 documentId: unpackedRef?.qid,
                 version: unpackedRef?.version,
                 blockRef: unpackedRef?.blockRef,
@@ -249,7 +226,7 @@ function EmbedControl({
         }
       }
     },
-    [block.props.ref],
+    [block.props.url],
   )
 
   return (
@@ -272,7 +249,7 @@ function EmbedControl({
           icon={<ExternalLink />}
           backgroundColor="$backgroundStrong"
           onPress={() => {
-            openUrl(block.props.ref, true)
+            openUrl(block.props.url, true)
           }}
         />
       </Tooltip>
@@ -358,226 +335,5 @@ function EmbedControl({
         </Popover>
       )}
     </XStack>
-  )
-}
-
-function EmbedForm({
-  block,
-  assign,
-  selected = false,
-}: {
-  block: Block<HMBlockSchema>
-  assign: any
-  editor: BlockNoteEditor<HMBlockSchema>
-  selected: boolean
-}) {
-  const [url, setUrl] = useState('')
-  const [loading, setLoading] = useState(false)
-  const [tabState, setTabState] = useState('embed')
-  const [error, setError] = useState<{
-    name: string
-    color: string | undefined
-  }>({
-    name: '',
-    color: undefined,
-  })
-  const theme = useTheme()
-  const {queryClient} = useAppContext()
-  const gwUrl = useGatewayUrlStream()
-  function submitEmbed(url: string) {
-    if (isPublicGatewayLink(url, gwUrl) || isHypermediaScheme(url)) {
-      const hmLink = normlizeHmId(url, gwUrl)
-      const ref = hmLink ? hmLink : url
-      assign({props: {ref: ref}} as EmbedType)
-    } else {
-      setLoading(true)
-      fetchWebLink(queryClient, url)
-        .then((res) => {
-          const fullHmId = hmIdWithVersion(
-            res?.hmId,
-            res?.hmVersion,
-            extractBlockRefOfUrl(url),
-          )
-          if (fullHmId) {
-            assign({props: {ref: fullHmId}} as EmbedType)
-          } else {
-            setError({
-              name: 'The provided url is not a hypermedia link',
-              color: 'red',
-            })
-          }
-          setLoading(false)
-        })
-        .catch((e) => {
-          setError({
-            name: 'The provided url is not a hypermedia link',
-            color: 'red',
-          })
-          setLoading(false)
-        })
-    }
-  }
-
-  return (
-    <YStack
-      //@ts-ignore
-      contentEditable={false}
-      position="relative"
-      borderWidth={0}
-      outlineWidth={0}
-    >
-      <Popover
-        placement="bottom"
-        size="$5"
-        defaultOpen={boolRegex.test(block.props.defaultOpen)}
-        stayInFrame
-      >
-        <Popover.Trigger asChild>
-          <Button
-            icon={<RiArticleLine fill={theme.color12.get()} />}
-            borderRadius={0}
-            size="$5"
-            justifyContent="flex-start"
-            backgroundColor="$color3"
-            hoverStyle={{
-              backgroundColor: '$color4',
-            }}
-          >
-            Add an Embed
-          </Button>
-        </Popover.Trigger>
-        <Popover.Content
-          padding={0}
-          elevation="$3"
-          overflow="hidden"
-          size="$5"
-          borderRadius="$5"
-          shadowColor="$shadowColor"
-          opacity={1}
-          enterStyle={{x: 0, y: -10, opacity: 0}}
-          exitStyle={{x: 0, y: -10, opacity: 0}}
-          animation={[
-            'fast',
-            {
-              opacity: {
-                overshootClamping: true,
-              },
-            },
-          ]}
-        >
-          <Tabs
-            value={tabState}
-            onValueChange={setTabState}
-            orientation="horizontal"
-            flexDirection="column"
-            width={500}
-            elevate
-          >
-            <Tabs.List
-              marginBottom="$-0.5"
-              backgroundColor="$background"
-              borderBottomColor="$color8"
-              borderBottomWidth="$1"
-              borderBottomLeftRadius={0}
-              borderBottomRightRadius={0}
-              borderRadius={0}
-            >
-              <Tabs.Tab
-                unstyled
-                value="embed"
-                paddingHorizontal="$4"
-                paddingVertical="$2"
-                borderBottomLeftRadius={0}
-                borderBottomRightRadius={0}
-                borderBottomWidth={'$1'}
-                hoverStyle={{
-                  backgroundColor: '$backgroundHover',
-                  cursor: 'pointer',
-                }}
-              >
-                <SizableText size="$2">Embed</SizableText>
-              </Tabs.Tab>
-            </Tabs.List>
-
-            <Tabs.Content value="embed">
-              <XStack
-                padding="$4"
-                alignItems="center"
-                backgroundColor="$background"
-              >
-                <Form
-                  alignItems="center"
-                  onSubmit={() => submitEmbed(url)}
-                  borderWidth={0}
-                >
-                  <YStack flex={1}>
-                    <XStack>
-                      <Input
-                        width={360}
-                        marginRight="$3"
-                        borderColor="$color8"
-                        borderWidth="$0.5"
-                        borderRadius="$3"
-                        size="$3.5"
-                        placeholder="Input embed link..."
-                        focusStyle={{
-                          borderColor: '$colorFocus',
-                          outlineWidth: 0,
-                        }}
-                        hoverStyle={{
-                          borderColor: '$colorFocus',
-                          outlineWidth: 0,
-                        }}
-                        onChange={(e) => {
-                          setUrl(e.nativeEvent.text)
-                          if (error.color)
-                            setError({
-                              name: '',
-                              color: undefined,
-                            })
-                        }}
-                        autoFocus={true}
-                      />
-                      <Form.Trigger asChild>
-                        <Button
-                          flex={0}
-                          flexShrink={0}
-                          borderRadius="$3"
-                          size="$3.5"
-                          theme={error.color === 'red' ? 'gray' : 'green'}
-                          disabled={error.color === 'red' ? true : false}
-                          focusStyle={{
-                            outlineWidth: 0,
-                          }}
-                        >
-                          {loading ? (
-                            <Spinner
-                              size="small"
-                              color="$green9"
-                              paddingHorizontal="$3"
-                            />
-                          ) : (
-                            'Embed'
-                          )}
-                        </Button>
-                      </Form.Trigger>
-                    </XStack>
-                    {error.name && (
-                      <SizableText
-                        size="$2"
-                        color={error.color}
-                        paddingTop="$2"
-                      >
-                        {error.name}
-                      </SizableText>
-                    )}
-                  </YStack>
-                </Form>
-              </XStack>
-            </Tabs.Content>
-          </Tabs>
-        </Popover.Content>
-      </Popover>
-    </YStack>
   )
 }
